Move WorkspaceScheduleForm
