+++ site/src/utils/schedule.test.ts
@@ -1,7 +1,5 @@
 import dayjs from "dayjs"
 import duration from "dayjs/plugin/duration"
-import { emptySchedule } from "pages/WorkspaceSettingsPage/WorkspaceSchedulePage/schedule"
-import { emptyTTL } from "pages/WorkspaceSettingsPage/WorkspaceSchedulePage/ttl"
 import { Workspace } from "../api/typesGenerated"
 import * as Mocks from "../testHelpers/entities"
 import {
@@ -12,7 +10,6 @@ import {
   getMaxDeadlineChange,
   getMinDeadline,
   stripTimezone,
-  scheduleChanged,
 } from "./schedule"
 
 dayjs.extend(duration)
@@ -75,73 +72,5 @@ describe("getMaxDeadlineChange", () => {
     const minDeadline = dayjs()
     // you can only subtract 2 hours even though the min is 2:40 less
     expect(getMaxDeadlineChange(deadline, minDeadline)).toEqual(2)
-  })
-})
-
-describe("scheduleChanged", () => {
-  describe("autostart", () => {
-    it("should be true if toggle values are different", () => {
-      const autostart = { autostartEnabled: true, ...emptySchedule }
-      const formValues = {
-        autostartEnabled: false,
-        ...emptySchedule,
-        autostopEnabled: false,
-        ttl: emptyTTL,
-      }
-      expect(scheduleChanged(autostart, formValues)).toBe(true)
-    })
-    it("should be true if schedule values are different", () => {
-      const autostart = { autostartEnabled: true, ...emptySchedule }
-      const formValues = {
-        autostartEnabled: true,
-        ...{ ...emptySchedule, monday: true, startTime: "09:00" },
-        autostopEnabled: false,
-        ttl: emptyTTL,
-      }
-      expect(scheduleChanged(autostart, formValues)).toBe(true)
-    })
-    it("should be false if all autostart values are the same", () => {
-      const autostart = { autostartEnabled: true, ...emptySchedule }
-      const formValues = {
-        autostartEnabled: true,
-        ...emptySchedule,
-        autostopEnabled: false,
-        ttl: emptyTTL,
-      }
-      expect(scheduleChanged(autostart, formValues)).toBe(false)
-    })
-  })
-
-  describe("autostop", () => {
-    it("should be true if toggle values are different", () => {
-      const autostop = { autostopEnabled: true, ttl: 1000 }
-      const formValues = {
-        autostartEnabled: false,
-        ...emptySchedule,
-        autostopEnabled: false,
-        ttl: 1000,
-      }
-      expect(scheduleChanged(autostop, formValues)).toBe(true)
-    })
-    it("should be true if ttl values are different", () => {
-      const autostop = { autostopEnabled: true, ttl: 1000 }
-      const formValues = {
-        autostartEnabled: false,
-        ...emptySchedule,
-        autostopEnabled: true,
-        ttl: 2000,
-      }
-      expect(scheduleChanged(autostop, formValues)).toBe(true)
-    })
-    it("should be false if all autostop values are the same", () => {
-      const autostop = { autostopEnabled: true, ttl: 1000 }
-      const formValues = {
-        autostartEnabled: false,
-        ...emptySchedule,
-        autostopEnabled: true,
-        ttl: 1000,
-      }
-      expect(scheduleChanged(autostop, formValues)).toBe(false)
-    })
   })
 })
